Working on a generate_data.py which will eventually be what generates the audio files that will be part of the golden dataset. Once this is complete, need to create another .py that will generate the 50 to 100 unique transcripts, and need to determine which values are worth keeping. Will need to create the respective json that is the correct output for each transcript in that .py.
Also created a dummy node js. Need to determine how I will broadcast this to a GUI front end.

When to involve FAST API? When to involve web sockets? How to set up a dummy twilio number to test it out? How to create an eval grade score based on the golden data set?

 All are questions I will need to address later on.